Tests split and fixed
App.enzyme.test.tsx renamed to Test.enzyme.test.tsx
