+++ src/App.test.tsx
@@ -1,8 +1,7 @@
-// Link.react.test.js
-import React from 'react';
+// App.test.js
+import React from 'react'
 import App from './App';
 import renderer from 'react-test-renderer'
-import waitForExpect from 'wait-for-expect'
 
 /**
  * Demo test showing how to use react-test-renderer
@@ -23,22 +22,4 @@ test('Link changes the class when hovered', () => {
   tree = component.toJSON()
   if(tree)
     expect(tree).toMatchSnapshot()
-})
-/**
- * Demo test showing how to use waitForExpect
- */
-test("it waits for the number to change", async () => {
-  let numberToChange = 10
-  // we are using random timeout here to simulate a real-time example
-  // of an async operation calling a callback at a non-deterministic time
-  const randomTimeout = Math.floor(Math.random() * 3000)
-  console.log(`Will be waiting for ${randomTimeout} msec`)
-
-  setTimeout(() => {
-    numberToChange = 100
-  }, randomTimeout)
-
-  await waitForExpect(() => {
-    expect(numberToChange).toEqual(100)
-  })
 })
